+++ cli/src/config.rs
@@ -480,6 +480,8 @@ pub struct WorkspaceConfig {
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub exclude: Vec<String>,
     #[serde(default, skip_serializing_if = "String::is_empty")]
+    pub idl: String,
+    #[serde(default, skip_serializing_if = "String::is_empty")]
     pub types: String,
 }